+++ SWP391_CareSkin_BE/DTOS/Requests/Product/ProductSearchRequestDTO.cs
@@ -5,8 +5,8 @@ namespace SWP391_CareSkin_BE.DTOS.Requests
         public string? Keyword { get; set; }
         public string? Category { get; set; }
         public int? BrandId { get; set; }
-        public double? MinPrice { get; set; }
-        public double? MaxPrice { get; set; }
+        public decimal? MinPrice { get; set; }
+        public decimal? MaxPrice { get; set; }
         public int? MinMl { get; set; }
         public int? MaxMl { get; set; }
         public string? SortBy { get; set; }
